Remove clang warning suppressions, and fix fallout
Summary: We were suppressing various clang warnings. Mostly harmless
fixes. I think the pdo one was a real buffer overrun waiting to happen.

Reviewed By: @edwinsmith, @bertmaher

Differential Revision: D1566241

+++ hphp/runtime/ext/fileinfo/libmagic/file.h
@@ -114,7 +114,7 @@
 #define  MAX(a,b)  (((a) > (b)) ? (a) : (b))
 #endif
 
-#define OFFSET_OOB(n, o, i)    ((n) < (o) || (i) > ((n) - (o)))
+#define OFFSET_OOB(n, o, i)    ((n) < (o) || (i) > ssize_t((n) - (o)))
 
 #ifndef HOWMANY
 # define HOWMANY (256 * 1024)  /* how much of the file to look at */
